+++ src/api/mypage.ts
@@ -27,3 +27,34 @@ export const submitReview = async (reservationId: number, reviewData: AddReviewD
   );
   return res;
 };
+
+// 마이페이지 > 내 정보 가져오기
+export const getMyInfo = async () => {
+  const res = await clientToken.get(`${import.meta.env.VITE_SERVER_URL}/user/mypage`);
+  return res.data;
+};
+
+// 마이페이지 > 내 정보 변경
+export const changeInfo = async (userName: string, phoneNum: string) => {
+  const res = await clientToken.put(`${import.meta.env.VITE_SERVER_URL}/user`, {
+    username: userName,
+    phonenumber: phoneNum,
+  });
+
+  return res.data;
+};
+
+// 마이페이지 > 비밀번호 변경
+export const changePw = async (oldPw: string, newPw: string) => {
+  const res = await clientToken.put(`${import.meta.env.VITE_SERVER_URL}/user/password`, {
+    oldPassword: oldPw,
+    newPassword: newPw,
+  });
+  return res.data;
+};
+
+// 마이페이지 > 회원탈퇴
+export const resignUser = async () => {
+  const res = await clientToken.delete(`${import.meta.env.VITE_SERVER_URL}/user`);
+  return res.data;
+};
